% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_date}
\alias{assert_date}
\title{assert a value can be cast to Date}
\usage{
assert_date(x)
}
\arguments{
\item{x}{the value to test}
}
\description{
assert a value can be cast to Date
}
\keyword{internal}
