% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_name}
\alias{assert_name}
\title{assert a value is a valid school holiday name}
\usage{
assert_name(x)
}
\arguments{
\item{x}{the value to test}
}
\description{
assert a value is a valid school holiday name
}
\keyword{internal}
