% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_year}
\alias{assert_year}
\title{assert a value can be considered as a year}
\usage{
assert_year(x)
}
\arguments{
\item{x}{the value to test}
}
\description{
assert a value can be considered as a year
}
\keyword{internal}
