% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_holidays.R
\name{bank_holiday}
\alias{bank_holiday}
\title{pick a bank holiday}
\usage{
bank_holiday(type)
}
\arguments{
\item{type}{one of the type of bank holidays, see \code{\link[=bank_holidays_names]{bank_holidays_names()}}}
}
\description{
Note this function returns a function with a \code{year} parameter.
}
