% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_holidays.R
\name{bank_holidays}
\alias{bank_holidays}
\alias{bh_paques}
\alias{bh_lundi_paques}
\alias{bh_vendredi_saint}
\alias{bh_ascension}
\alias{bh_pentecote}
\alias{bh_jour_an}
\alias{bh_premier_mai}
\alias{bh_fete_travail}
\alias{bh_huit_mai}
\alias{bh_victoire_allies}
\alias{bh_quatorze_juillet}
\alias{bh_fete_nationale}
\alias{bh_quinze_aout}
\alias{bh_assomption}
\alias{bh_toussaint}
\alias{bh_onze_novembre}
\alias{bh_armistice}
\alias{bh_noel}
\alias{bh_saint_etienne}
\title{bank holidays}
\usage{
bh_paques(year)

bh_lundi_paques(year)

bh_vendredi_saint(year)

bh_ascension(year)

bh_pentecote(year)

bh_jour_an(year)

bh_premier_mai(year)

bh_fete_travail(year)

bh_huit_mai(year)

bh_victoire_allies(year)

bh_quatorze_juillet(year)

bh_fete_nationale(year)

bh_quinze_aout(year)

bh_assomption(year)

bh_toussaint(year)

bh_onze_novembre(year)

bh_armistice(year)

bh_noel(year)

bh_saint_etienne(year)
}
\arguments{
\item{year}{an integer}
}
\description{
bank holidays
}
\keyword{internal}
