% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{bank_holidays_names}
\alias{bank_holidays_names}
\alias{bank_holidays_names_unique}
\title{available bank holidays names}
\usage{
bank_holidays_names()

bank_holidays_names_unique()
}
\description{
available bank holidays names
}
