% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{build_zone_string}
\alias{build_zone_string}
\title{build the zone string}
\usage{
build_zone_string(zone)
}
\arguments{
\item{zone}{a character, either "A", "B" or "C"}
}
\description{
build the zone string
}
\keyword{internal}
