% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{days_until_holidays}
\alias{days_until_holidays}
\title{how many days until the next holidays?}
\usage{
days_until_holidays(zone = NULL, quiet = FALSE)
}
\arguments{
\item{zone}{a character, either "A", "B" or "C"}

\item{quiet}{whether or not to return a message}
}
\value{
the number of days before the next holidays
}
\description{
how many days until the next holidays in France?
Check for your zone using the \code{zone} param.
}
\examples{
days_until_holidays()
}
