% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{full_holidays_names}
\alias{full_holidays_names}
\alias{short_holidays_names}
\title{available holidays names}
\usage{
full_holidays_names()

short_holidays_names()
}
\description{
available holidays names
}
