% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{get_holidays}
\alias{get_holidays}
\title{get all holidays for a year}
\usage{
get_holidays(year, zone = NULL, name = NULL, drop_zones = FALSE)
}
\arguments{
\item{year}{a vector of integers between 2008 and 2018}

\item{zone}{a character, either "A", "B" or "C"}

\item{name}{a character, the name of the holiday period}

\item{drop_zones}{a logical, if TRUE and \code{zone} is not NULL, return only the specified zones}
}
\description{
Note that you can use short names instead of full names (see \code{\link[=holidays_names_full]{holidays_names_full()}})
}
