% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_holidays.R
\encoding{UTF-8}
\name{get_year_bank_holidays}
\alias{get_year_bank_holidays}
\title{get all bank holidays for a year}
\usage{
get_year_bank_holidays(year, include_alsace = FALSE)
}
\arguments{
\item{year}{an integer}

\item{include_alsace}{a logical, if TRUE, include the bank holidays
from the Alsace-Moselle départements}
}
\description{
See \code{\link[=bank_holidays]{bank_holidays()}} for individual functions.
}
