% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{holidays_names_full}
\alias{holidays_names_full}
\alias{holidays_names_short}
\title{available holidays names}
\usage{
holidays_names_full()

holidays_names_short()
}
\description{
available holidays names
}
