% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_holidays.R
\name{is_bank_holiday}
\alias{is_bank_holiday}
\title{check if a date is a bank holiday}
\usage{
is_bank_holiday(d)
}
\arguments{
\item{d}{a value that can be cast to a Date}
}
\description{
check if a date is a bank holiday
}
