% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{is_holiday}
\alias{is_holiday}
\title{check if a date is a holiday}
\usage{
is_holiday(d, zone = NULL)
}
\arguments{
\item{d}{a value that can be cast to a Date}

\item{zone}{a character, either "A", "B" or "C"}
}
\description{
check if a date is a holiday
}
