% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{next_holidays}
\alias{next_holidays}
\title{when will we be on holidays?}
\usage{
next_holidays(zone = NULL)
}
\arguments{
\item{zone}{a character, either "A", "B" or "C"}
}
\value{
The date of the next holidays
}
\description{
When will we be on holidays in France?
Check for your zone using the \code{zone} param.
}
\examples{
next_holidays(zone = "C")
}
